# openapi.yaml
openapi: 3.0.0
info:
  title: Token API 审计与管理系统
  description: |
    基于 Django 的 AI API 代理审计与管理系统，支持多厂商 AI 模型的统一管理、用户配额控制、成本计算和使用审计。
    
    ## 认证方式
    
    本API支持两种认证方式：
    
    1. **JWT Token认证** - 用于管理后台操作
       - 在 Authorization 头中使用: `Bearer <jwt_token>`
    
    2. **API Key认证** - 用于AI服务调用
       - 在 Authorization 头中使用: `Bearer sk-audit-{32位字符}`
    
    ## 权限说明
    
    - **超级管理员**: 可以访问所有 `/api/admin/` 开头的接口
    - **普通用户**: 只能访问自己的资料和配额信息
    - **API Key**: 仅用于 `/v1/` 下的 OpenAI 兼容接口
  version: 1.0.0
  contact:
    name: API Support
    email: admin@example.com
servers:
  - url: http://127.0.0.1:8000
    description: 开发环境
  - url: https://your-domain.com
    description: 生产环境
security:
  - JWTAuth: []
  - APIKeyAuth: []
paths:
...
  /api/admin/quotas/{id}/:
    get:
      tags:
        - 管理员 - 配额管理
      summary: 获取配额详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 配额详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserQuota'
    put:
      tags:
        - 管理员 - 配额管理
      summary: 更新配额
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuotaUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserQuota'
    delete:
      tags:
        - 管理员 - 配额管理
      summary: 删除配额
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 删除成功
...
  /v1/chat/completions:
    post:
      tags:
        - OpenAI兼容接口
      summary: 聊天完成
      description: |
        OpenAI兼容的聊天完成接口，支持多种AI模型。
        使用API Key认证：Authorization: Bearer sk-audit-{32位字符}
      security:
        - APIKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          description: 聊天完成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
        '400':
          description: 请求参数错误
        '401':
          description: 认证失败
        '403':
          description: 模型不在配额范围内
        '429':
          description: 配额超限
...
  /v1/usage:
    get:
      tags:
        - OpenAI兼容接口
      summary: 查看使用情况
      description: 获取当前API Key的配额使用情况
      security:
        - APIKeyAuth: []
      responses:
        '200':
          description: 使用情况
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageResponse'
components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token认证，用于管理后台操作
    APIKeyAuth:
      type: http
      scheme: bearer
      description: API Key认证，格式：Bearer sk-audit-{32位字符}
  schemas:
...
    UserQuota:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        user:
          type: integer
          description: 用户ID
        user_name:
          type: string
          readOnly: true
          description: 用户名
        user_email:
          type: string
          readOnly: true
          description: 用户邮箱
        model_group:
          type: integer
          description: 模型组ID
        model_group_name:
          type: string
          readOnly: true
          description: 模型组名称
        api_key:
          type: string
          readOnly: true
          description: API密钥
        masked_api_key:
          type: string
          readOnly: true
          description: 掩码后的API密钥
        total_quota:
          type: number
          format: decimal
          description: 总配额($)
        used_quota:
          type: number
          format: decimal
          description: 已使用配额($)
        remaining_quota:
          type: number
          format: decimal
          readOnly: true
          description: 剩余配额($)
        usage_percentage:
          type: number
          readOnly: true
          description: 使用百分比
        period_type:
          type: string
          enum: [monthly, weekly, daily, unlimited]
          description: 配额周期类型
        period_start:
          type: string
          format: date-time
          description: 周期开始时间
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: 过期时间
        rate_limit_per_minute:
          type: integer
          description: 每分钟请求限制
        rate_limit_per_hour:
          type: integer
          description: 每小时请求限制
        rate_limit_per_day:
          type: integer
          description: 每日请求限制
        is_active:
          type: boolean
          description: 是否激活
        auto_renew:
          type: boolean
          description: 是否自动续费
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
...
    QuotaUpdateRequest:
      type: object
      properties:
        total_quota:
          type: number
          format: decimal
        period_type:
          type: string
          enum: [monthly, weekly, daily, unlimited]
        expires_at:
          type: string
          format: date-time
          nullable: true
        rate_limit_per_minute:
          type: integer
        rate_limit_per_hour:
          type: integer
        rate_limit_per_day:
          type: integer
        is_active:
          type: boolean
        auto_renew:
          type: boolean

    # API请求记录
...
    ChatCompletionRequest:
      type: object
      required:
        - model
        - messages
      properties:
        model:
          type: string
          description: 要使用的模型名称
          example: gpt-4o
        messages:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [system, user, assistant]
              content:
                type: string
          description: 对话消息列表
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 1
          description: 采样温度
        max_tokens:
          type: integer
          description: 最大生成tokens数
        stream:
          type: boolean
          default: false
          description: 是否流式返回
    ChatCompletionResponse:
      type: object
      properties:
        id:
          type: string
          description: 聊天完成ID
        object:
          type: string
          enum: [chat.completion]
        created:
          type: integer
          description: 创建时间戳
        model:
          type: string
          description: 使用的模型
        choices:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              message:
                type: object
                properties:
                  role:
                    type: string
                  content:
                    type: string
              finish_reason:
                type: string
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer
...
    UsageResponse:
      type: object
      properties:
        quota:
          type: object
          properties:
            total_quota:
              type: number
              description: 总配额
            used_quota:
              type: number
              description: 已使用配额
            remaining_quota:
              type: number
              description: 剩余配额
            model_group:
              type: string
              description: 模型组名称
            period_type:
              type: string
              description: 周期类型
            expires_at:
              type: string
              format: date-time
              nullable: true
              description: 过期时间
        recent_requests:
          type: integer
          description: 最近请求数
        recent_cost:
          type: number
          description: 最近花费

    # 分页响应
...